back/pyproject.toml
[project]
name = "ai-reader-backend"
version = "0.1.0"
description = "AI Reader Backend API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "openai>=1.0.0",
    "newspaper3k>=0.2.8",
    "lxml_html_clean>=0.1.0",
    "nltk>=3.8.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
